refactor: reorganize pages
