+++ frontend-clean/pages/join/friends.js
